Split NodeGraphEditor
